Add Home Login and Register functionality
Add code for home to the home link in base html,
add code for login to login link in base html, added
view for login in views file and url changed code in login
html template to accommodate same.  Added code for signup
to the signup link in the base html page, added view for signup
and also added url.  In creating these views fixed bugs in
realtion to No reverse match and page not found errors