Fix placeholder image and no of likes functionalit
Added link for placeholder to postdetail.html. Correct authentication
of no of likes on postdetail.  postread added placeholder image.
Removed image field from the card element.  Get "please login to
like or comment on a post" to display when user is not logged in.
split code between two lines in pagination number code.